# public/openapi.yaml
openapi: 3.0.3
info:
  title: OfferGuide API
  version: 2.0.0
  description: >
    OfferGuide (sanjeeda.io career services portal) backend contract.


    Sprint 8 completed this file. Sprints 1–7 shipped 34 paths but only
    Sprint 5's own 13 were ever documented here, which made Story 8.1.1
    ("verify every endpoint in Swagger") impossible to execute. Every route
    that exists in `src/app/api/offerguide/` is now declared, with its auth
    mode, status codes and response shape.


    ## Auth modes

    Four modes, named exactly as the Sprint 8 handoff §3.1 table names them:


    * **PUBLIC** — no cookie, no token. Reference data only.

    * **guestOrAuth** — a `guestToken` httpOnly cookie **or** a portal JWT.
      Behaviour is identical either way, scoped to whichever identity is
      present; a JWT wins when both are sent. Never returns one candidate's
      data to another.

    * **bearerAuth** — a portal JWT only. A guest cookie is rejected.

    * **adminAuth** — a portal JWT whose account's stored `role` is `admin`.
      Rejects no credentials, a guest cookie, and an ordinary registered-user
      JWT alike. The role is read from the database on every check, so a
      demotion takes effect immediately rather than at the next token expiry.

    ### Swagger UI limitation (read before verifying)

    The **Authorize** button can only drive the JWT cases. It cannot attach the
    `guestToken` httpOnly cookie. Every `guestOrAuth` operation must therefore
...

...


...

...

...

...

...
servers:
...
security: []
paths:
...
  /api/offerguide/candidate-profile:
    get:
      tags: [Candidate Profile]
      summary: Get the caller's candidate profile
      description: guestOrAuth. Scoped to the caller's own identity.
      security: [{ BearerAuth: [] }, { GuestCookie: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CandidateProfile" }
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Candidate Profile]
      summary: Create the caller's candidate profile
      description: >
        guestOrAuth. With no identity present, mints a `guestToken` and sets
        it as an httpOnly cookie on the response — this is the operation that
        makes a guest a guest.
      security: [{ BearerAuth: [] }, { GuestCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CandidateProfile" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CandidateProfile" }
        "400":
          $ref: "#/components/responses/BadRequest"
    patch:
      tags: [Candidate Profile]
      summary: Update the caller's candidate profile
      description: >
        guestOrAuth. Partial merge — only the fields sent are changed. A
        return visit updates the existing profile rather than creating a
        second one.
      security: [{ BearerAuth: [] }, { GuestCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CandidateProfile" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CandidateProfile" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        The portal's own JWT. On a `guestOrAuth` route it takes precedence
        over the guest cookie when both are present.
    GuestCookie:
      type: apiKey
      in: cookie
      name: guestToken
      description: >
        httpOnly cookie minted by `POST /candidate-profile` when no identity
        is present. **Swagger UI cannot attach this** — verify guest-mode
        behaviour outside Swagger with `curl -b` or a browser session.
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        The portal's own JWT, belonging to an account whose stored `role` is
        `admin` (Sprint 9). The same credential as `BearerAuth` — what separates
        the two is the account behind the token, not anything the client sends.

        The role is read from `sanjeedausers` on every check rather than taken
        from a token claim, so a demotion applies immediately instead of at the
        next expiry. Authorization is expressed as the `portal.admin.access`
        permission, resolved through the static role→permission map in
        `src/lib/portal/permissions.ts`; no call site compares the role directly.

        Fails closed. No token, an expired token, a token for a deleted account,
        an unrecognised role, or a database error all resolve to the public tier,
        which does not hold `portal.admin.access`.

        Sprint 9 replaced an interim env-configured header gate here. It was
        removed entirely rather than kept as a fallback — there is no second way
        into these routes.
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: >
        Not found — also returned when the resource belongs to another
        candidate, so the contract never confirms that someone else's data
        exists.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }

    # ----------------------------------------------------------- candidate
    ConsentSettings:
      type: object
      description: >
        Candidate-level, stored on `candidate_profiles.consentSettings`.
        Every toggle defaults to **Off** — privacy-first. When
        `shareAnonymous` is false, every selection is treated as Off
        regardless of its stored value, and the sub-toggles are dimmed in the
        UI. Framing is anonymised throughout: no copy implies personal data is
        shared.
      properties:
        shareAnonymous:
          type: boolean
          default: false
          description: The master toggle, `consent_share_anonymous`.
        selections:
          type: object
          additionalProperties: { type: boolean }
          description: >
            Keyed by `toggleId`. Validated against `OgConsentToggles`, so an
            unknown or retired id is rejected with 400.
    CandidateProfile:
      type: object
      description: >
        SCR-001. 24 fields, of which only `careerStage` and
        `preferredWorkArrangement` are required. One profile per identity;
        a return visit updates it rather than creating a second.
      properties:
        id: { type: integer, readOnly: true }
        userInfoId:
          type: integer
          nullable: true
          description: Null while the candidate is a guest.
        guestToken:
          type: string
          nullable: true
          readOnly: true
        careerStage: { type: string }
        preferredWorkArrangement: { type: string }
        currentCountry: { type: string, nullable: true }
        currentCity: { type: string, nullable: true }
        consentSettings: { $ref: "#/components/schemas/ConsentSettings" }
      additionalProperties: true

    # --------------------------------------------------------------- session
...